config naming strategy of jpa, modify get cart by customer and create cart detail logic

+++ BE_Project/src/main/java/com/example/be_project/DTO/CartDetailDTO.java
@@ -14,6 +14,6 @@ public class CartDetailDTO {
     private int bookId;
     private String bookTitle;
     private String bookImage;
-    private Long bookPrice;
+    private double bookPrice;
     private int amount;
 }
